fix compilation against libnl-3.x
Add configuration to compile against 3.1 and 3.2 without breaking 3.0
That's a bit of mess with version on .pc file and the part which belong or not to the shared lib name.

libnl-genl-3.1 is broken and can't be used

Starting from 3.2.0, I could have used libnl-genl-3.0.pc file like in patch V1 and not hardcode adding libnl-genl-3 into LIBS
I consider this a matter of tast for now, coded like previous versions that can't use that way to code.

Added an echo to display against which libnl version iw is compiled.
That particulary help me to test with various versions because of the mess on .pc file version as libnl>=3.2.0 has again pc file version with 3.0 number.

Removed the version_check part instead of extending blindly for the new versions.
That was mostly dead code because prior check on NLLIBNAME not empty already care.
The only case where that could matter would be using something stupid like
NLLIBNAME=foo make

Compilation tested to work against
- libnl-2.0
- libnl-3.0
- libnl-3.1
- libnl-3.2.0
- libnl-3.2.2

Tested to fail without a libnl .pc file

Signed-off-by: Gilles Espinasse <g.esp@free.fr>