iw: add set_mcast_rate command support
This patch adds the support to run the set_mcast_rate() command on
adhoc and mesh_point vifs. With this command it is possible to tune
the bitrate to use when sending group frames. This command can be used
even if the vifs has already joint the ibss/mesh network.

Signed-off-by: Antonio Quartulli <ordex@autistici.org>